requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
httpx==0.28.1
pyyaml==6.0.2
pydantic==2.10.4
jinja2==3.1.5
openai==1.58.1
